working with command list from npm
