+++ package.json
@@ -44,7 +44,7 @@
     "ejs": "^2.5.7",
     "execa": "^0.8.0",
     "fs-extra": "^5.0.0",
-    "globby": "^7.1.1",
+    "globby": "^6.0.0",
     "inquirer": "^4.0.2",
     "pkg-dir": "^2.0.0",
     "typings-core": "^2.3.3",
@@ -59,7 +59,7 @@
     "@types/execa": "^0.8.1",
     "@types/fs-extra": "^5.0.0",
     "@types/glob": "^5.0.30",
-    "@types/globby": "^6.1.0",
+    "@types/globby": "^0.6.0",
     "@types/grunt": "~0.4.0",
     "@types/inquirer": "0.0.36",
     "@types/mockery": "^1.4.29",
