+++ src/allCommands.ts
@@ -2,20 +2,13 @@ import {
 	loadCommands,
 	enumerateInstalledCommands,
 	enumerateBuiltInCommands,
-	LoadedCommands,
-	YargsCommandNames,
-	isEjected,
-	setDefaultGroup
+	LoadedCommands
 } from './loadCommands';
 import {
 	initCommandLoader,
-	createBuiltInCommandLoader,
-	CommandsMap,
-	CommandWrapper }
+	createBuiltInCommandLoader }
 	from './command';
 import config from './config';
-import * as chalk from 'chalk';
-import { NpmPackageDetails } from './interfaces';
 
 const commands: LoadedCommands = {
 	commandsMap: new Map(),
@@ -42,61 +35,16 @@ export async function loadBuiltInCommands(): Promise<LoadedCommands> {
 	return await loadCommands(builtInCommandsPaths, builtInCommandLoader);
 }
 
-export function createInstallableCommandPrompts(availableCommands: NpmPackageDetails[]): LoadedCommands {
-	const commandsMap: CommandsMap = new Map();
-	const yargsCommandNames: YargsCommandNames = new Map();
-	const regEx = /@dojo\/cli-([^-]+)-(.+)/;
-
-	availableCommands.forEach((command) => {
-		const [ , group, name ] = regEx.exec(command.name) as string[];
-		const compositeKey = `${group}-${name}`;
-		const installCommand = `npm i ${command.name}`;
-		const commandWrapper: CommandWrapper = {
-			name,
-			group,
-			path: installCommand,
-			description: command.description,
-			register: () => {},
-			run: () => {
-				console.log(`\nTo install this command run ${chalk.green(installCommand)}\n`);
-				return Promise.resolve();
-			}
-		};
-
-		if (!isEjected(group, name)) {
-			if (!commandsMap.has(group)) {
-				setDefaultGroup(commandsMap, group, commandWrapper);
-				yargsCommandNames.set(group, new Set());
-			}
-
-			if (!commandsMap.has(compositeKey)) {
-				commandsMap.set(compositeKey, commandWrapper);
-			}
-
-			const groupCommandNames = yargsCommandNames.get(group);
-			if (groupCommandNames) {
-				groupCommandNames.add(compositeKey);
-			}
-		}
-	});
-
-	return {
-		commandsMap,
-		yargsCommandNames
-	};
-}
-
-export default async function loadAllCommands(availableCommands: NpmPackageDetails[]): Promise<LoadedCommands> {
+export default async function loadAllCommands(): Promise<LoadedCommands> {
 	if (loaded) {
 		return Promise.resolve(commands);
 	}
 
-	const installableCommandPrompts = createInstallableCommandPrompts(availableCommands);
 	const builtInCommands = await loadBuiltInCommands();
 	const installedCommands = await loadExternalCommands();
 
-	commands.commandsMap = new Map([...installableCommandPrompts.commandsMap, ...installedCommands.commandsMap, ...builtInCommands.commandsMap]);
-	commands.yargsCommandNames = new Map([...installableCommandPrompts.yargsCommandNames, ...installedCommands.yargsCommandNames, ...builtInCommands.yargsCommandNames]);
+	commands.commandsMap = new Map([...installedCommands.commandsMap, ...builtInCommands.commandsMap]);
+	commands.yargsCommandNames = new Map([...installedCommands.yargsCommandNames, ...builtInCommands.yargsCommandNames]);
 	loaded = true;
 	return Promise.resolve(commands);
 }
