+++ src/index.ts
@@ -5,7 +5,7 @@ import updateNotifier from './updateNotifier';
 import registerCommands from './registerCommands';
 import { join } from 'path';
 import commandLoader from './allCommands';
-import installableCommands from './installableCommands';
+import installableCommands, { mergeInstalledCommandsWithAvailableCommands } from './installableCommands';
 const pkgDir = require('pkg-dir');
 
 /**
@@ -23,9 +23,10 @@ async function init() {
 		updateNotifier(packageJson, 0);
 
 		const availableCommands = await installableCommands(packageJson.name);
-		const allCommands = await commandLoader(availableCommands);
+		const allCommands = await commandLoader();
+		const mergedCommands = mergeInstalledCommandsWithAvailableCommands(allCommands, availableCommands);
 
-		registerCommands(yargs, allCommands.commandsMap, allCommands.yargsCommandNames);
+		registerCommands(yargs, mergedCommands.commandsMap, mergedCommands.yargsCommandNames);
 	} catch (err) {
 		console.log(`Commands are not available: ${err}`);
 	}
